requirements.txt
mysql-connector-python>=8.0.25
pytest>=6.2.4
